feat: converts files to typescript (and required changes that comes with the transformation)
